+++ lib/index.js
@@ -1,66 +1,111 @@
 'use strict';
 
-// module.exports = {};
-// библиотека
-
 const fs = require('fs');
-const TWO_WEEKS = 14 * 24 * 60 * 60 * 1000;
-const LOG_LEVELS = ['LOG', 'INFO', 'WARN', 'ERROR'];
+const path = require('path');
 
 const padNumber = (number) => String(number).padStart(2, '0');
-const METHODS = ['getHours', 'getMinutes', 'getSeconds'];
 
-const getCurrentTime = (now = new Date()) =>
-  METHODS.map((current) => padNumber(now[current]())).join(':');
-
-const formatMessage = (message) => {
-  const timestamp = getCurrentTime();
-  return `${timestamp} ${message}`;
+const getCurrentTime = () => {
+  const now = new Date();
+  const hours = padNumber(now.getHours());
+  const minutes = padNumber(now.getMinutes());
+  const seconds = padNumber(now.getSeconds());
+  return `${hours}:${minutes}:${seconds}`;
 };
-const getLogFileName = () => {
+
+const formatMessage = (level, message) => {
+  const timestamp = getCurrentTime();
+  return `${timestamp} [${level}] ${message}`;
+};
+const deleteLogFile = (logFile) => {
+  fs.unlink(logFile, (err) => {
+    if (err) {
+      console.error(`Ошибка при удалении файла логов: ${err}`);
+    } else {
+      console.info(`Файл логов ${logFile} успешно удален.`);
+    }
+  });
+};
+const setLogExpirationTimer = (logFilePath, keepDays) => {
+  const deleteDate = new Date();
+  deleteDate.setDate(deleteDate.getDate() - keepDays);
+
+  const checkDelete = () => {
+    if (fs.existsSync(logFilePath)) {
+      const fileDate = new Date(fs.statSync(logFilePath).ctime);
+      if (fileDate <= deleteDate) {
+        deleteLogFile(logFilePath);
+      }
+    }
+  };
+
+  setInterval(checkDelete, 60 * 60 * 1000);
+};
+const getLogFilePath = (config) => {
   const currentDate = new Date();
   const year = currentDate.getFullYear();
   const month = padNumber(currentDate.getMonth() + 1);
   const day = padNumber(currentDate.getDate());
-  return `${day}-${month}-${year}.log`;
+  const fileName = config.json ? `${year}-${month}-${day}.json` : `${year}-${month}-${day}.log`;
+  return path.join(config.path, fileName);
 };
-class Logger {
-  logFile = null;
-  stream = null;
-  time = null;
+const createLogFile = (config) => {
+  const logFilePath = getLogFilePath(config);
 
-  writeLog(message, consoleMessage) {
-    this.stream.write(`${message} \n`);
-    process.stdout.write(`${consoleMessage} \n`);
+  if (!fs.existsSync(config.path)) {
+    fs.mkdirSync(config.path, { recursive: true });
   }
 
-  deleteLogFile() {
-    fs.unlink(this.logFile, (err) => {
-      err
-        ? this.error(`Ошибка при удалении логов ${err}`)
-        : this.log(`Файл логов ${this.logFile} успешно удален`);
-    });
-  }
-  setLogExpirationTimer(duration) {
-    this.timer ? clearTimeout(this.timer) : '';
-    this.timer = setTimeout(() => void this.deleteLogFile(), duration);
-  }
-  createLogfile() {
-    this.logFile = getLogFileName();
-    this.stream = fs.createWriteStream(this.logFile, { flags: 'a' });
+  const stream = fs.createWriteStream(logFilePath, {
+    flags: 'a',
+    defaultEncoding: 'utf8',
+    highWaterMark: config.writeBuffer,
+  });
 
-    this.setLogExpirationTimer(TWO_WEEKS);
-  }
+  setLogExpirationTimer(logFilePath, config.keepDays);
 
-  info(level, message) {
-    if (!LOG_LEVELS.includes(level)) return;
-    const formattedMessage = formatMessage(level, message);
-    this.writeLog(formattedMessage, message);
-  }
+  return stream;
+};
 
-  log = this.info.bind(this, 'LOG');
-  warn = this.info.bind(this, 'WARN');
-  error = this.info.bind(this, 'ERROR');
-}
+const logger = (stream, levels, config) => {
+  let lastLogTime = new Date().getTime();
 
-module.exports = Logger;
+  const logFunctions = {};
+
+  levels.forEach((level) => {
+    logFunctions[level.toLowerCase()] = (message) => {
+      const currentTime = new Date().getTime();
+      const timeElapsed = currentTime - lastLogTime;
+      const delay = Math.max(0, config.writeInterval - timeElapsed);
+      lastLogTime = currentTime + delay;
+
+      setTimeout(() => {
+        const formattedMessage = formatMessage(level, message);
+        stream.write(`${formattedMessage}\n`);
+        process.stdout.write(`${formattedMessage}\n`);
+      }, delay);
+    };
+  });
+
+  return logFunctions;
+};
+
+const createLogger = (config) => {
+  const logFileStream = createLogFile(config);
+  return logger(logFileStream, ['LOG', 'INFO', 'WARN', 'ERROR'], config);
+};
+
+const opts = {
+  path: './log',
+  writeInterval: 3000,
+  writeBuffer: 64 * 1024,
+  keepDays: 5,
+  json: false,
+};
+
+const log = createLogger(opts);
+
+log.log('Пример лога');
+log.info('Пример информационного лога');
+log.warn('Пример предупреждения');
+log.error('Пример ошибки');
